TIP-842: Move family searchable repository to Akeneo\Pim\Structure (enrich bundle)
